+++ Emerald/src/postprocessing/postprocessing_blur_gaussian.cc
@@ -761,7 +761,7 @@ end:
 /** Please see header for spec */
 PUBLIC RENDERING_CONTEXT_CALL EMERALD_API void postprocessing_blur_gaussian_execute(     __notnull postprocessing_blur_gaussian            blur,
                                                                                     __in           unsigned int                            n_taps,
-                                                                                    __in           unsigned int                            n_iterations,
+                                                                                    __in           float                                   n_iterations,
                                                                                     __in __notnull ogl_texture                             src_texture,
                                                                                     __in           postprocessing_blur_gaussian_resolution blur_resolution)
 {
@@ -954,7 +954,7 @@ PUBLIC RENDERING_CONTEXT_CALL EMERALD_API void postprocessing_blur_gaussian_exec
 
     /* Step 2): Kick off! */
     for (unsigned int n_iteration = 0;
-                      n_iteration < n_iterations;
+                      n_iteration < (unsigned int) n_iterations;
                     ++n_iteration)
     {
         /* Horizontal pass */
